// src/app/_services/swagger-auto-generated/model/responseDataSetSingerDto.ts
/**
 * API for controlling church music
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.2
 * Contact: kb.developer.br@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ResponseError } from './responseError';
import { SingerDto } from './singerDto';

export interface ResponseDataSetSingerDto { 
    content?: Array<SingerDto>;
    error?: ResponseError;
}